modified the CodeBlock
